Fixing these infinite recursing methods
Oops, that was silly. Thanks @t00thpick1

+++ src/main/java/com/gmail/nossr50/api/ExperienceAPI.java
@@ -50,7 +50,7 @@ public final class ExperienceAPI {
 
     @Deprecated
     public static void addRawXP(Player player, String skillType, int XP) {
-        addRawXP(player, skillType, XP);
+        addRawXP(player, skillType, (float) XP);
     }
 
     /**
@@ -70,7 +70,7 @@ public final class ExperienceAPI {
 
     @Deprecated
     public static void addRawXPOffline(String playerName, String skillType, int XP) {
-        addRawXPOffline(playerName, skillType, XP);
+        addRawXPOffline(playerName, skillType, (float) XP);
     }
 
     /**
